refactor(directory): ♻ convert SegmentNumber and Primariness to shorts

+++ services/directory/src/FilterLists.Directory.Infrastructure.Migrations/Migrations/20200828001720_Initial.cs
@@ -276,8 +276,8 @@ namespace FilterLists.Directory.Infrastructure.Migrations.Migrations
                 {
                     Id = table.Column<int>(nullable: false)
                         .Annotation("Npgsql:ValueGenerationStrategy", NpgsqlValueGenerationStrategy.IdentityByDefaultColumn),
-                    SegmentNumber = table.Column<int>(nullable: false, defaultValue: 1),
-                    Primariness = table.Column<int>(nullable: false, defaultValue: 1),
+                    SegmentNumber = table.Column<short>(nullable: false, defaultValue: (short)1),
+                    Primariness = table.Column<short>(nullable: false, defaultValue: (short)1),
                     Url = table.Column<string>(nullable: false),
                     FilterListId = table.Column<int>(nullable: false)
                 },
