+++ services/directory/src/FilterLists.Directory.Infrastructure/Persistence/Queries/Entities/FilterListViewUrl.cs
@@ -4,8 +4,8 @@ namespace FilterLists.Directory.Infrastructure.Persistence.Queries.Entities
 {
     public class FilterListViewUrl
     {
-        public int SegmentNumber { get; set; }
-        public int Primariness { get; set; }
+        public short SegmentNumber { get; set; }
+        public short Primariness { get; set; }
         public Uri Url { get; set; } = null!;
     }
 }
